Rename command and use click for options
