+++ dali_mon
@@ -1,4 +1,4 @@
 #!/bin/bash
 source ./env/bin/activate
 pip3 install -r requirements.txt > /dev/null
-python3 source/main.py --version $*
+python3 source/dali_mon.py $*
