+++ tests/requirements.txt
@@ -2,4 +2,5 @@ pytest
 coverage
 pyserial
 pyusb
+click
 termcolor
